Commit update UserService, add @Operation, @Schema, @ApiResponse.

+++ src/main/java/com/tms/service/UserService.java
@@ -1,5 +1,7 @@
 package com.tms.service;
 
+import com.tms.exception.SecurityCredentialsForbiddenException;
+import com.tms.exception.UserInfoNotFoundException;
 import com.tms.models.Role;
 import com.tms.models.UserInfo;
 import com.tms.repository.UserRepository;
@@ -8,6 +10,7 @@ import com.tms.security.domain.SecurityCredentials;
 import com.tms.security.repository.SecurityCredentialsRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.dao.OptimisticLockingFailureException;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
@@ -28,10 +31,12 @@ public class UserService {
     public Optional<UserInfo> getUserById(Integer id) {
         String login = SecurityContextHolder.getContext().getAuthentication().getName();
         Optional<SecurityCredentials> securityCredentialsOptional = securityCredentialsRepository.findByUserLogin(login);
-        if (securityCredentialsOptional.isPresent()){
+        if (securityCredentialsOptional.isPresent()) {
             SecurityCredentials securityCredentials = securityCredentialsOptional.get();
-            if (id.equals(securityCredentials.getUserId()) || (securityCredentials.getUserRole() == Role.ADMIN)){
+            if (id.equals(securityCredentials.getUserId()) || (securityCredentials.getUserRole() == Role.ADMIN)) {
                 return userRepository.findById(id);
+            }else {
+                throw new UserInfoNotFoundException();
             }
         }
         return Optional.empty();
@@ -43,22 +48,58 @@ public class UserService {
     }
 
     public void addFavoritesFile(RequestParametersId requestParametersId) {
-        log.info("Saving new favorites file {}", requestParametersId);
-        userRepository.addFileToUser(requestParametersId.getUserId(), requestParametersId.getFileId());
+        String login = SecurityContextHolder.getContext().getAuthentication().getName();
+        Optional<SecurityCredentials> securityCredentialsOptional = securityCredentialsRepository.findByUserLogin(login);
+        if (securityCredentialsOptional.isPresent()) {
+            SecurityCredentials securityCredentials = securityCredentialsOptional.get();
+            if (requestParametersId.getUserId().equals(securityCredentials.getUserId()) || (securityCredentials.getUserRole() == Role.ADMIN)) {
+                log.info("Saving new favorites file {}", requestParametersId);
+                userRepository.addFileToUser(requestParametersId.getUserId(), requestParametersId.getFileId());
+            } else {
+                throw new SecurityCredentialsForbiddenException();
+            }
+        }
     }
 
     public void updateUser(UserInfo userInfo) {
-        log.info("Update user {}", userInfo);
-        userRepository.saveAndFlush(userInfo);
+        String login = SecurityContextHolder.getContext().getAuthentication().getName();
+        Optional<SecurityCredentials> securityCredentialsOptional = securityCredentialsRepository.findByUserLogin(login);
+        if (securityCredentialsOptional.isPresent()) {
+            SecurityCredentials securityCredentials = securityCredentialsOptional.get();
+            if (userInfo.getId().equals(securityCredentials.getUserId()) || (securityCredentials.getUserRole() == Role.ADMIN)) {
+                log.info("Update user {}", userInfo);
+                userRepository.saveAndFlush(userInfo);
+            } else {
+                throw new SecurityCredentialsForbiddenException();
+            }
+        }
     }
 
     public void deleteUserById(Integer id) {
-        log.info("Delete user {}", id);
-        userRepository.deleteById(id);
+        String login = SecurityContextHolder.getContext().getAuthentication().getName();
+        Optional<SecurityCredentials> securityCredentialsOptional = securityCredentialsRepository.findByUserLogin(login);
+        if (securityCredentialsOptional.isPresent()) {
+            SecurityCredentials securityCredentials = securityCredentialsOptional.get();
+            if (id.equals(securityCredentials.getUserId()) || (securityCredentials.getUserRole() == Role.ADMIN)) {
+                log.info("Delete user {}", id);
+                userRepository.deleteById(id);
+            } else {
+                throw new SecurityCredentialsForbiddenException();
+            }
+        }
     }
 
     public void deleteByFavoritesFile(RequestParametersId requestParametersId) {
-        log.info("Delete favorites file {}", requestParametersId);
-        userRepository.deleteByFavoritesFile(requestParametersId.getUserId(), requestParametersId.getFileId());
+        String login = SecurityContextHolder.getContext().getAuthentication().getName();
+        Optional<SecurityCredentials> securityCredentialsOptional = securityCredentialsRepository.findByUserLogin(login);
+        if (securityCredentialsOptional.isPresent()) {
+            SecurityCredentials securityCredentials = securityCredentialsOptional.get();
+            if (requestParametersId.getUserId().equals(securityCredentials.getUserId()) || (securityCredentials.getUserRole() == Role.ADMIN)) {
+                log.info("Delete favorites file {}", requestParametersId);
+                userRepository.deleteByFavoritesFile(requestParametersId.getUserId(), requestParametersId.getFileId());
+            } else {
+                throw new SecurityCredentialsForbiddenException();
+            }
+        }
     }
 }
